% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catastrozip2st.R
\name{catastrozip2st}
\alias{catastrozip2st}
\title{catastrozip2st}
\usage{
catastrozip2st(ziplink, tempdir = "./temp/", name, Hfloor = 3)
}
\arguments{
\item{ziplink}{--> this is the zip url for a single municipality (captured in catastroprovince)}

\item{tempdir}{--> this where temporary files are stores (zips, unzips...)}

\item{name}{--> this is the name of the muninipality (captured in catastroprovince)}

\item{Hfloor}{--> asign a height of each floor above the ground (stimation of building height)}
}
\value{
list of 3 elemenst (building, buildingpart, otherconstruction) in simple feature collection.
This parts will be merged and unified in a gpkg with the catastroprovince() function.
}
\description{
# catastrozip2st downloads a zip link into a temporal folder and extracts:
    - building.gml
    - buildingpart.gml
    - otherconstruction.gml
    It returns a list of 3 elements, each in st format
    Note1: it includes the calculation of Hmax heigth for buildingparts
}
\examples{
\dontrun{
ziplink <- 'http://www.catastro.minhap.es/INSPIRE/Buildings/56/56101-MELILLA/A.ES.SDGC.BU.56101.zip'
name = 'MELILLA'
catastrozip2st(ziplink,tempdir,name)
}
}
