% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvinceLinks.R
\name{ProvinceLinks}
\alias{ProvinceLinks}
\title{ProvinceLinks}
\usage{
ProvinceLinks(catastropush)
}
\arguments{
\item{catastropush}{--> given an atom url returns a data.frame Provices|URL}
}
\value{
data frame with 2 columns (Provincia & URL)
}
\description{
Given a catastropush it returns a data frame with the url .xml for each Province
}
\examples{
\dontrun{
ProvinceLinks("http://www.catastro.minhap.es/INSPIRE/buildings/ES.SDGC.bu.atom.xml")
}
}
